+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Introduce new administrative channel management endpoints `/admin/v1/realms/{realmId}/channels`.
+- Introduce new administrative subscription management endpoints `/admin/v1/realms/{realmId}/subscriptions`.
 
 ### Changed
 - Add support for upgraded verneMQ broker. This requires the acceptance of their [EULA](https://vernemq.com/end-user-license-agreement/),